+++ customers/customers_service.go
@@ -3,5 +3,5 @@ package customers
 import "context"
 
 type CustomerService interface {
-	GetCustomerById(ctx context.Context, id int) (*Customer, error)
+	GetCustomerById(ctx context.Context, id int64) (*Customer, error)
 }
